Fix CONNECT bug and do minor refactoring
- Added \r\n to CONNECT generator. This previously caused an error on
connect which in turn terminated the client.
- Started working on reporting errors and validating input better. How
this should be designed is still under consideration and I am not really
sure what is the best way to go. I think the best way to go is by
granular error reporting in the API. In that way the user can `.map {}` on
result types if they don't care about them or they could `switch` on
particular errors..